requirements.txt
requests
colorama
PyYAML